+++ viking.js
@@ -2810,6 +2810,250 @@ Viking.View.Helpers.textField = function (model, attribute, options) {
 // TODO: url_field
 // TODO: week_field
 // TODO: year_field
+// urlFor(modelOrUrl, options)
+// ===========================
+//
+// Returns the URL for the model. Note that by default +:onlyPath+ is +true+ so
+// you'll get the relative "/controller/action" instead of the fully qualified
+// URL like "http://example.com/controller/action".
+//
+// Passing the model will trigger the named route for that record. The lookup
+// will happen on the name of the class. So passing a Workshop object will attempt
+// to use the +workshopPath+ route. If you have a nested route, such as
+// +adminWorkshopPath+ you'll have to call that explicitly (it's impossible
+// for +urlFor+ to guess that route).
+//
+// If a string is passed instead of a model, the string is simply returned.
+//
+// ==== Options
+//   - anchor: Specifies the anchor name to be appended to the path.
+//   - onlyPath: If +true+, returns the relative URL (omitting the protocol,
+//               host name, and port) (+false+ by default).
+//   - trailingSlash: If true, adds a trailing slash, as in "/archive/2005/".
+//                     Note that this is currently not recommended since it
+//                     breaks caching.
+//   - host: Overrides the default (current) host if provided.
+//   - port: Optionally specify the port to connect to
+//   - protocol: Overrides the default (current) protocol if provided.
+//   - scriptName: Specifies application path relative to domain root. If
+//                 provided, prepends application path
+//   - user: Inline HTTP authentication (only plucked out if +:password+ is also present).
+//   - password: Inline HTTP authentication (only plucked out if +:user+ is also present).
+//
+// ==== Examples
+//   urlFor(Workshop)
+//   // uses +workshopsUrl()+
+//   # => http://www.example.com/workshops
+//
+//   urlFor(new Workshop())
+//   // relies on Workshop answering isNew() (and in this case returning true)
+//   # => http://www.example.com/workshops
+//
+//   urlFor(workshop)
+//   // uses +workshopUrl(model)+
+//   // which calls workshop.toParam() which by default returns the id
+//   // => http://www.example.com/workshops/5
+//
+//   // toParam() can be re-defined in a model to provide different URL names:
+//   // => http://www.example.com/workshops/1-workshop-name
+//
+//   urlFor(workshop, {anchor: 'location'})
+//   // => http://www.example.com/workshops/5#location
+//
+//   urlFor(workshop, {onlyPath: true})
+//   // => /workshops/5
+//
+//   urlFor(workshop, {trailingSlash: true})
+//   // => http://www.example.com/workshops/5/
+//
+//   urlFor(workshop, {host: 'myhost'})
+//   // => http://myhost.com/workshops/5
+//
+//   urlFor(workshop, {port: 9292})
+//   // => http://www.example.com:9292/workshops/5
+//
+//   urlFor(workshop, {protocol: 'https'})
+//   // => https://www.example.com/workshops/5
+//
+//   urlFor(workshop, {scriptName: '/location'})
+//   // => http://example.com/location/workshops/5
+//
+//   urlFor(workshop, {user: 'username', password: 'password'})
+//   // => http://username:password@example.com/workshops/5
+//
+//   urlFor("http://www.example.com")
+//   // => http://www.example.com
+//
+// TODO: support polymorhic_paths... ie [blog, post] => /blogs/1/posts/3
+//       polymorphic_url([blog, post]) # => "http://example.com/blogs/1/posts/1"
+//       polymorphic_url([:admin, blog, post]) # => "http://example.com/admin/blogs/1/posts/1"
+//       polymorphic_url([user, :blog, post]) # => "http://example.com/users/1/blog/posts/1"
+//       polymorphic_url(Comment) # => "http://example.com/comments"
+//
+urlFor = function (modelOrUrl, options) {
+    if (typeof modelOrUrl === 'string') {
+        return modelOrUrl;
+    }
+    
+    options = _.extend({
+        onlyPath: false,
+        trailingSlash: false,
+        host: window.location.hostname,
+        port: window.location.port,
+        scriptName: '',
+        protocol: window.location.protocol.replace(':','')
+    }, options);
+    
+    var route;
+    var klass = modelOrUrl.modelName.camelize().constantize();
+    if (modelOrUrl instanceof klass) {
+        if (modelOrUrl.isNew()) {
+            route = (klass.modelName.pluralize() + (options.onlyPath ? 'Path' : 'Url')).constantize();
+            route = route();
+        } else {
+            route = (klass.modelName + (options.onlyPath ? 'Path' : 'Url')).constantize();
+            route = route(modelOrUrl);
+        }
+    } else {
+        route = (modelOrUrl.modelName.pluralize() + (options.onlyPath ? 'Path' : 'Url')).constantize();
+        route = route();
+    }
+    
+    if (!options.onlyPath) {
+        route = options.protocol + '://' + options.host + (options.port ? ':' : '') + options.port + options.scriptName + route;
+        
+        if (options.user && options.password) {
+            route = route.replace('://', '://' + options.user + ':' + options.password + '@');
+        }
+    }
+    
+    if (options.trailingSlash) {
+        route += '/';
+    }
+    
+    if (options.anchor) {
+        route += '#' + options.anchor;
+    }
+    
+    return route;
+};
+// linkTo(modelOrUrl, name, options)
+// =================================
+//
+// Creates a link tag to the +modelOrUrl+. If +modelOrUrl+ is a model a href is
+// generated from the routes. If a String is passed it is used as the href. See
+// the documentation for +urlFor+ for how +modelOrUrl+ is translated into a url.
+//
+// ==== Signatures
+//
+//   linkTo(url)
+//   
+//   linkTo(url, body)
+//
+//   linkTo(url, body, options)
+//
+//   linkTo(url, options)
+//
+//   linkTo(url, options, body)
+//
+//
+// ==== Options
+//   - data: This option can be used to add custom data attributes.
+// TODO: method not supported yet
+//   - method: an HTTP verb. This modifier will dynamically create an HTML form
+//             and immediately submit the form for processing using the HTTP verb
+//             specified. Useful for having links perform a POST operation
+//             in dangerous actions like deleting a record (which search bots can
+//             follow while spidering your site). Supported verbs are +post+,
+//             +:delete+, +:patch+, and +:put+.
+//
+// ==== Examples
+//
+//   linkTo(profilePath(profile), "Profile")
+//   // => <a href="/profiles/1">Profile</a>
+//
+// or the even pithier
+//
+//   linkTo(profile, "Profile")
+//   // => <a href="/profiles/1">Profile</a>
+//
+// Similarly,
+//
+//   linkTo(profiles_path(), "Profiles")
+//   // => <a href="/profiles">Profiles</a>
+//
+//
+// You can use a function as well if your link target is hard to fit into the
+// name parameter. EJS example:
+//
+//   <%= linkTo(profile, function () { %>
+//     <strong><%= profile.name %></strong> -- <span>Check it out!</span> 
+//   <% }) %>
+//   // => <a href="/profiles/1">
+//          <strong>David</strong> -- <span>Check it out!</span>
+//         </a>
+//
+// Classes and ids for CSS are easy to produce:
+//
+//   linkTo(articles_path(), "Articles", {id: 'news', class: 'article'})
+//   // => <a href="/articles" class="article" id="news">Articles</a>
+//
+// +linkTo+ can also produce links with anchors or query strings:
+//
+//   linkTo(profilePath(profile, {anchor: "wall"}), "Comment wall")
+//   // => <a href="/profiles/1#wall">Comment wall</a>
+//
+//   linkTo(searches_path({foo: "bar", baz: "quux"}), "Nonsense search")
+//   // => <a href="/searches?foo=bar&amp;baz=quux">Nonsense search</a>
+//
+// TODO: method not supported yet
+// The only option specific to +linkTo+ (+:method+) is used as follows:
+//
+//   linkTo("http://www.example.com", "Destroy", {method: "delete"})
+//   // => <a href='http://www.example.com' rel="nofollow" data-method="delete">Destroy</a>
+//
+// You can also use custom data attributes using the +:data+ option:
+//
+//   linkTo("http://www.rubyonrails.org/", "Visit Other Site", { data: { confirm: "Are you sure?" }})
+//   // => <a href="http://www.rubyonrails.org/" data-confirm="Are you sure?">Visit Other Site</a>
+Viking.View.Helpers.linkTo = function (modelOrUrl, name, options) {
+    if (typeof name === 'object') {
+        var tmp = options;
+        options = name;
+        name = tmp;
+    }
+    
+    options = _.extend({
+        href: urlFor(modelOrUrl)
+    }, options);
+    
+    if (!name) { 
+        if (modelOrUrl instanceof Viking.Model) {
+            if (modelOrUrl.isNew()) {
+                name = 'new ' + modelOrUrl.modelName.humanize();
+            } else {
+                name = modelOrUrl.modelName.humanize() + ' #' + modelOrUrl.id;
+            }
+        } else if (modelOrUrl.modelName) {
+            name = modelOrUrl.modelName.pluralize().humanize();
+        } else {
+            name = modelOrUrl;
+        }
+    }
+
+    return Viking.View.Helpers.contentTag('a', name, options);
+};
+
+        // 
+        // html_options, options, name = options, name, block if block_given?
+        // options ||= {}
+        // 
+        // html_options = convert_options_to_data_attributes(options, html_options)
+        // 
+        // url = url_for(options)
+        // html_options['href'] ||= url
+        // 
+        // content_tag(:a, name || url, html_options, &block)
 // mailTo(email, name, options)
 // mailTo(email, contentFunc)
 // mailTo(email, options, contentFunc)
@@ -2863,9 +3107,7 @@ Viking.View.Helpers.mailTo = function (email, name, options) {
     
     // handle (email, name, options), (email, contentFunc), and 
     // (email, options, contentFunc) formats
-    if (typeof name === 'function') {
-        name = name;
-    } else if (typeof options === 'function') {
+    if (typeof options === 'function') {
         tmp = options;
         options = name;
         name = tmp;
@@ -2888,6 +3130,9 @@ Viking.View.Helpers.mailTo = function (email, name, options) {
 
     return Viking.View.Helpers.contentTag('a', name, options, false);
 };
+
+
+
 function FormBuilder(model, options) {
     options || (options = {});